+++ src/layer/x86/mish_x86.cpp
@@ -117,7 +117,7 @@ int Mish_x86::forward_inplace(Mat& bottom_top_blob, const Option& opt) const
 #endif // __SSE2__
         for (; i < size; i++)
         {
-            *ptr = *ptr * tanh(log(exp(*ptr) + 1.f));
+            *ptr = *ptr * tanhf(logf(expf(*ptr) + 1.f));
             ptr++;
         }
     }
